Move fixture client entity setup to fixtures
