+++ docs/PlaywrightStandards.md
@@ -13,22 +13,43 @@ playwright/
 ├── package.json
 ├── playwright.config.ts
 ├── fixtures/
+│   ├── client/
+│   │   ├── create_client.ts
+│   │   ├── build_minimal_client_payload.ts
+│   │   └── minimal.json
 │   └── example-entity/
 │       ├── example_entity_fixture.ts
 │       └── minimal.json
 ├── tests/
 │   └── example.spec.ts
 └── utils/
-    └── example_common_code.ts
+    ├── random_text.ts
+    └── sirius_api.ts
 ```
 
 - `playwright/package.json`: Playwright-specific scripts and dependencies.
 - `playwright/playwright.config.ts`: Central Playwright configuration (projects, reporters, retries, timeouts, and test discovery).
 - `playwright/tests/`: Test specs (`*.spec.ts`).
-- `playwright/fixtures/`: Shared test data and fixture setup helpers.
-- `playwright/utils/`: Reusable helper functions used by tests and fixtures.
+- `playwright/fixtures/`: Shared test data, entity factories, API-backed setup helpers, and fixture setup code used to prepare test state for specific domains.
+- `playwright/utils/`: Reusable generic helper functions used by tests and fixtures.
 
 Keep tests feature-focused, and prefer extracting reusable setup/data logic into `fixtures/` and `utils/` to avoid duplication.
+
+### Choosing between `fixtures/` and `utils/`
+
+- Use `playwright/fixtures/` for code that prepares test state for a specific domain, such as:
+  - creating entities through the API
+  - building default payloads for a specific entity
+  - domain-specific setup helpers
+  - fixture data files such as `minimal.json`
+
+- Use `playwright/utils/` for code that is generic and cross-cutting, such as:
+  - API transport wrappers
+  - random data generators
+  - wait/polling helpers
+  - generic formatting or date helpers
+
+If a helper is specific to a single entity, colocate it under that entity's fixture directory rather than `utils/`.
 
 ## Standards
